requirements.txt
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
pillow==10.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
